requirements.txt
langchain
langchain-community 
langchain-core 
langchain-ollama 
langgraph 
faiss-cpu 
shap
pypdf
sentence-transformers
streamlit
crewai
python-dotenv
litellm
google-genai
langchain-google-genai
langchain-experimental 
pandas
pymupdf 
requests
crewai-tools
fpdf